requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126

torch
torchvision
transformers == 5.0.0rc2
diffusers == 0.36.0
accelerate == 1.12.0
numpy == 2.3.5
sentencepiece == 0.2.1
datasets == 4.4.1
protobuf == 6.33.2  # T5 依赖
pillow == 12.0.0
tqdm == 4.67.1
pennylane
matplotlib